feat: Update writeStream
This commit updates the `writeStream` code in the `challenges/writeStream` directory. It includes changes to the `axios` and `https` modules, as well as the `fs` and `crypto` modules. Additionally, the `package.json` file has been updated to include dependencies for `axios` and `ws`. The `.gitignore` file has been modified to include `package-lock.json`.